Initial commit
It won't work on anyone else's PC and I can't upload the music files for copyright reasons. I made it over a year ago and I clearly didn't know how to do some things, but in hindsight this is a pretty cool project! Saving it for tracking my own progress